Some changes in the pull request
- Delete the flag isScoreVisible
- Add new JSON only for Guesses
